+++ src/model.ml
@@ -340,6 +340,7 @@ type ('id, 'term) mterm_node  =
   | Msource
   | Mselfaddress
   | Mchainid
+  | Mmetadata
   (* variable *)
   | Mvar              of 'id * 'term var_kind_gen * temp * delta
   (* rational *)
@@ -1303,6 +1304,7 @@ let cmp_mterm_node
     | Msource, Msource                                                                 -> true
     | Mselfaddress, Mselfaddress                                                       -> true
     | Mchainid, Mchainid                                                               -> true
+    | Mmetadata, Mmetadata                                                             -> true
     (* variable *)
     | Mvar (id1, k1, t1, d1), Mvar (id2, k2, t2, d2)                                   -> cmpi id1 id2 && cmp_var_kind k1 k2 && cmp_temp t1 t2 && cmp_delta d1 d2
     (* rational *)
@@ -1663,8 +1665,9 @@ let map_term_node_internal (fi : ident -> ident) (g : 'id -> 'id) (ft : type_ ->
   | Mcaller                        -> Mcaller
   | Mbalance                       -> Mbalance
   | Msource                        -> Msource
-  | Mchainid                       -> Mchainid
   | Mselfaddress                   -> Mselfaddress
+  | Mchainid                       -> Mchainid
+  | Mmetadata                      -> Mmetadata
   (* variable *)
   | Mvar (id, k, t, d)             -> Mvar (g id, map_var_kind f k, map_temp fi t, map_delta d)
   (* rational *)
@@ -2030,6 +2033,7 @@ let fold_term (f : 'a -> ('id mterm_gen) -> 'a) (accu : 'a) (term : 'id mterm_ge
   | Msource                               -> accu
   | Mselfaddress                          -> accu
   | Mchainid                              -> accu
+  | Mmetadata                             -> accu
   (* variable *)
   | Mvar (_, k, _, _)                     -> fold_var_kind f accu k
   (* rational *)
@@ -2863,6 +2867,9 @@ let fold_map_term
 
   | Mchainid ->
     g Mchainid, accu
+
+  | Mmetadata ->
+    g Mmetadata, accu
 
 
   (* variable *)
@@ -5020,6 +5027,7 @@ end = struct
       | Msource                    -> "source"::accu
       | Mselfaddress               -> "selfaddress"::accu
       | Mchainid                   -> "chainid"::accu
+      | Mmetadata                  -> "metadata"::accu
 
       | Mget (an, ck, _)
       | Mselect (an, ck, _, _, _)